requirements.txt
openjdk==17.*
python==3.9.*
pandas==2.2.*
bbmap==39.01
bowtie2==2.5.1
bwa==0.7.17
bwa-mem2==2.2.1
fastp==0.23.2
fastqc==0.11.9
gatk4==4.6.1.0
picard==3.*
qualimap==2.2.2d
samtools==1.20.*
somatic-sniper==1.0.5.0
bcftools==1.16
tqdm
varscan==2.4.4
yaml
black
numpy
pillow
pygraphviz
snakefmt
snakemake==7.18.*
elprep==5.1.3
vardict-java==1.8.3
snpeff==5.1
perl-dbi
perl-lwp-simple
conda
msisensor-pro
genefuse
shortuuid
pyranges
upsetplot
seaborn
scikit-learn
matplotlib-venn
celery
redis-py
docker-py
python-dotenv